create user_batches_new table: a clone of user_batches table, but compatible with MySQL 8.0.28; use it instead of user_batches everywhere

+++ testhelpers/db.go
@@ -248,6 +248,9 @@ func emptyDB(db *sql.DB, dbName string) error {
 			_ = tx.Rollback()
 			return scanErr
 		}
+		if tableName == dbName+".user_batches" { // skip this broken table
+			continue
+		}
 		// DELETE is MUCH faster than TRUNCATE on empty tables
 		_, err = tx.Exec("DELETE FROM " + tableName)
 		if err != nil {
